chainstate: Introduce the `chainstate-launcher` crate
Used to create chainstate together with storage based on a config.
