+++ chainstate/Cargo.toml
@@ -4,8 +4,6 @@ license = "MIT"
 name = "chainstate"
 version = "0.1.0"
 
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
 [dependencies]
 chainstate-storage = {path = './storage'}
 chainstate-types = {path = './types'}
@@ -15,9 +13,6 @@ crypto = {path = '../crypto'}
 logging = {path = '../logging'}
 rpc = {path = '../rpc'}
 serialization = {path = "../serialization"}
-storage = {path = "../storage"}
-storage-inmemory = {path = "../storage/inmemory"}
-storage-lmdb = {path = "../storage/lmdb"}
 subsystem = {path = '../subsystem'}
 utxo = {path = '../utxo'}
 utils = {path = '../utils'}
