+++ chainstate/src/config.rs
@@ -56,45 +56,3 @@ impl Default for ChainstateConfig {
         }
     }
 }
-
-/// Storage type to use
-#[derive(Serialize, Deserialize, Debug, Clone, PartialEq, Eq)]
-pub enum StorageBackend {
-    #[serde(rename = "lmdb")]
-    Lmdb,
-    #[serde(rename = "inmemory", alias = "in-memory")]
-    InMemory,
-}
-
-impl std::str::FromStr for StorageBackend {
-    type Err = serde::de::value::Error;
-
-    fn from_str(s: &str) -> Result<Self, Self::Err> {
-        let de = serde::de::value::StrDeserializer::new(s);
-        serde::Deserialize::deserialize(de)
-    }
-}
-
-impl Default for StorageBackend {
-    fn default() -> Self {
-        Self::Lmdb
-    }
-}
-
-/// Storage configuration
-#[derive(Serialize, Deserialize, Debug, Default)]
-pub struct ChainstateAndStorageConfig {
-    /// Storage backend to use
-    #[serde(default)]
-    pub storage_backend: StorageBackend,
-
-    /// Chainstate configuration
-    #[serde(flatten)]
-    pub chainstate_config: ChainstateConfig,
-}
-
-impl ChainstateAndStorageConfig {
-    pub fn new() -> Self {
-        Self::default()
-    }
-}
