+++ node/Cargo.toml
@@ -10,7 +10,8 @@ license = "MIT"
 # Local dependencies
 blockprod = { path = "../blockprod/" }
 common = { path = "../common/" }
-chainstate = { path = "../chainstate/" }
+chainstate = { path = "../chainstate" }
+chainstate-launcher = { path = "../chainstate/launcher" }
 logging = { path = "../logging/" }
 mempool = { path = "../mempool/" }
 p2p = { path = "../p2p/" }
